major updates
bug fix
added lookahead compression
switched to S_32 format for input
fixed buffer overflow in clipper.

+++ alsa_pipe/alsa_pipe.h
@@ -20,7 +20,7 @@ int setup_alsa_pipe(char* recording_iface,
 //CHANNEL2-4
 
 //the data buffer should be the same size as set in setup
-int get_audio(short* data,int bsize);
+int get_audio(int* data,int bsize);
 int queue_audio(int* data);
 //puts in the next segment to be played
